Visualize colors in HTML widget

Display a grid of color swatches with optional names or color codes.

Usage

visual_colors(
  colors,
  names = NULL,
  num_per_row = 30,
  title = NULL,
  label_mode = c("auto", "chinese", "pinyin", "rgb", "hex")
)

Arguments

colors

A character vector of hex color codes.

names

Optional. A character vector of names for each color. Default is NULL, which means hex color codes will be displayed. You can pass any labels (e.g., RGB values, custom names) via this parameter.

num_per_row

Number of colors per row. Default is 30.

title

Optional title for the visualization. Default is NULL.

label_mode

Label layout mode. Default is "auto", which detects the label type from names. Use "chinese", "pinyin", "rgb", or "hex" to force a layout optimized for that label style.

Value

An HTML widget.

Examples

# Visualize a simple color palette
visual_colors(
  colors = c("#FF0000", "#00FF00", "#0000FF"),
  names = c("Red", "Green", "Blue")
)

  

          

            
Red

          

        
        	
          

            
Green

          

        
        	
          

            
Blue

          

        
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
      

visual_colors(
  colors = c("#FF0000", "#00FF00"),
  names = c("(255, 0, 0)", "(0, 255, 0)")
)

  

          

            
(255, 0, 0)

          

        
        	
          

            
(0, 255, 0)

          

        
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
      

visual_colors(thisplot::palette_list$Paired)

  

          

            
#A6CEE3

          

        
        	
          

            
#1F78B4

          

        
        	
          

            
#B2DF8A

          

        
        	
          

            
#33A02C

          

        
        	
          

            
#FDBF6F

          

        
        	
          

            
#FF7F00

          

        
        	
          

            
#FB9A99

          

        
        	
          

            
#E31A1C

          

        
        	
          

            
#CAB2D6

          

        
        	
          

            
#6A3D9A

          

        
        	
          

            
#FFFF99

          

        
        	
          

            
#B15928

          

        
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
        	
      

# Use with ChineseColors
cc <- ChineseColors()
visual_colors(
  colors = cc$blue[1:60],
  title = "Chinese Blue Colors"
)

  
Chinese Blue Colors

  

          

            
#D2F0F4

          

        
        	
          

            
#C6E6E8

          

        
        	
          

            
#B0D5DF

          

        
        	
          

            
#A2D2E2

          

        
        	
          

            
#85C0DB

          

        
        	
          

            
#8ABCD1

          

        
        	
          

            
#8FB2C9

          

        
        	
          

            
#93B5CF

          

        
        	
          

            
#8AABCC

          

        
        	
          

            
#8BA3C7

          

        
        	
          

            
#88ABDA

          

        
        	
          

            
#6E9BC5

          

        
        	
          

            
#619AC3

          

        
        	
          

            
#5698C3

          

        
        	
          

            
#4994C4

          

        
        	
          

            
#2486B9

          

        
        	
          

            
#1781B5

          

        
        	
          

            
#2983BB

          

        
        	
          

            
#1177B0

          

        
        	
          

            
#1677B3

          

        
        	
          

            
#2B73AF

          

        
        	
          

            
#3271AE

          

        
        	
          

            
#2474B5

          

        
        	
          

            
#2376B7

          

        
        	
          

            
#2177B8

          

        
        	
          

            
#1772B4

          

        
        	
          

            
#126BAE

          

        
        	
          

            
#1661AB

          

        
        	
          

            
#0F59A4

          

        
        	
          

            
#15559A

          

        
      

          

            
#2A5390

          

        
        	
          

            
#2E4E7E

          

        
        	
          

            
#22406A

          

        
        	
          

            
#06436F

          

        
        	
          

            
#144A74

          

        
        	
          

            
#12507B

          

        
        	
          

            
#354E6B

          

        
        	
          

            
#45465E

          

        
        	
          

            
#454659

          

        
        	
          

            
#535164

          

        
        	
          

            
#6B798E

          

        
        	
          

            
#5E7987

          

        
        	
          

            
#4182A4

          

        
        	
          

            
#4E7CA1

          

        
        	
          

            
#346C9C

          

        
        	
          

            
#3170A7

          

        
        	
          

            
#11659A

          

        
        	
          

            
#106898

          

        
        	
          

            
#158BB8

          

        
        	
          

            
#2F90B9

          

        
        	
          

            
#1A94BC

          

        
        	
          

            
#22A2C3

          

        
        	
          

            
#5CB3CC

          

        
        	
          

            
#63BBD0

          

        
        	
          

            
#3CBEE7

          

        
        	
          

            
#66A9C9

          

        
        	
          

            
#A3BBDB

          

        
        	
          

            
#AED0EE

          

        
        	
          

            
#BCD4E7

          

        
        	
          

            
#BACCD9